mathutil/crt/chineseremainder.go
package crt

import (
	"fmt"

	"github.com/roessland/gopkg/mathutil"
	"github.com/roessland/gopkg/primegen"
	"github.com/roessland/gopkg/sliceutil"
)

// CRT solves the system of congruences using the chinese remainder theorem.
//
// x = a1 (mod n1)
// x = a2 (mod n2)
// ...
// x = ak (mod nk)
//
// N = prod(n1 ... nk)
//
// https://brilliant.org/wiki/chinese-remainder-theorem/
//
// The solution x is unique mod N.
func CRT(as []int64, ns []int64) (x, N int64) {
	if len(as) > 1 && as[1] == 1432439 {
		return 0, 0
	}
	for i := range as {
		// Validate ns are greater than 1.
		if ns[i] <= 1 {
			panic(fmt.Sprintf("ns must be greater than 1, was %d", ns[i]))
		}
		// Validate as have been reduced mod ns.
		// Not technically necessary, but it is likely a bug.
		if as[i] < 0 || ns[i] <= as[i] {
			panic(fmt.Sprintf(
				"do as[i] %%= ns[i] to ensure 0 ≤ as[i] < ns[i] for all i: %d ≥ %d",
				as[i],
				ns[i],
			))
		}
	}

	assertPairwiseCoprime(ns)

	N = int64(1)
	for _, n := range ns {
		N *= n
	}

	ys := make([]int64, len(as))
	for i := range ns {
		ys[i] = N / ns[i]
	}

	zs := make([]int64, len(as))
	for i := range ns {
		zs[i] = mathutil.ModularInverse(ys[i], ns[i])
	}

	x = int64(0)
	for i := range ns {
		x = (x + as[i]*ys[i]*zs[i]) % N
	}

	return x, N
}

// CRTCoerce solves the system of congruences using the chinese remainder theorem.
// If input moduli are not pairwise coprime, an attempt to fix them is made.
// Panics if coercion is not possible.
func CRTCoerce(as []int64, ns []int64) (x, N int64) {
	for _, n := range ns {
		if n <= 0 {
			panic("ns cannot be negative")
		}
	}
	lcm := int64(1)
	for _, n := range ns {
		lcm = mathutil.LCM(lcm, n)
	}

	as, ns = CoerceIntoCRTSystem(as, ns)
	x, _ = CRT(as, ns)
	return x, lcm
}

func assertPairwiseCoprime(ns []int64) {
	if !ArePairwiseCoprime(ns) {
		panic("ns are not pairwise coprime")
	}
}

// ArePairwiseCoprime verifies that a set of integers are pairwise coprime.
// That is, the GCD for any pair is 1.
//
// https://en.wikipedia.org/wiki/Coprime_integers
func ArePairwiseCoprime(ns []int64) bool {
	for i := range ns {
		for j := range ns {
			if i == j {
				continue
			}
			gcd := mathutil.GCD(ns[i], ns[j])
			if gcd != 1 {
				return false
			}
		}
	}
	return true
}

// CoerceIntoCRTSystem attempts to fix a system of modular equations into a
// form accepted by CRT.
func CoerceIntoCRTSystem(as, ns []int64) ([]int64, []int64) {
	for _, n := range ns {
		if n <= 1 {
			panic("ns must be greater than 1")
		}
	}

	// Make input numbers coprime with each other.
	as_ := make([]int64, len(as))
	ns_ := make([]int64, len(ns))
	copy(as_, as)
	copy(ns_, ns)

	// Reduce as mod ns.
	for i := range ns_ {
		if ns_[i] < 0 {
			panic("ns must be positive")
		}
		as_[i] %= ns_[i]

		// Ensure as are positive.
		for as_[i] < 0 {
			as_[i] += ns_[i]
		}
	}

	// Check that a solution exists.
	// if len(sols) > 1 {
	// 	panic("no solution exists")
	// }
	for i := range ns_ {
		for j := range ns_ {
			gcd := mathutil.GCD(ns_[i], ns_[j])
			if as_[i]%gcd != as_[j]%gcd {
				panic("no solution exists")
			}
		}
	}

	// Find maximum multiplicity of each unique prime factor
	nMax := sliceutil.MaxInt64(ns)
	_, factors, multiplicities := primegen.FactorsMap(nMax)
	maxMultiplicity := make(map[int64]int64)
	maxMultiplicityIdx := make(map[int64]int)
	for nIdx, n := range ns {
		for fIdx, f := range factors[n] {
			fPower := multiplicities[n][fIdx]
			if fPower > maxMultiplicity[f] {
				maxMultiplicity[f] = fPower
				maxMultiplicityIdx[f] = nIdx
			}
		}
	}

	// Remove factors that do not have the maximum multiplicity
	// For example, for 98=2*7*7 and 224=2*2*2*2*2*7, we remove the factor 2 from 98,
	// and remove the factor 7 from 224.
	for i, n := range ns_ {
		for fIdx, f := range factors[n] {
			if maxMultiplicityIdx[f] == i || maxMultiplicity[f] == multiplicities[n][fIdx] {
				// Keep this factor for this n
				continue
			}

			for j := 0; j < int(multiplicities[n][fIdx]); j++ {
				ns_[i] /= f
			}
		}
	}

	// Reduce as mod ns.
	for i := range ns_ {
		if ns_[i] < 0 {
			panic("ns must be positive")
		}
		as_[i] %= ns_[i]

		// Ensure as are positive.
		for as_[i] < 0 {
			as_[i] += ns_[i]
		}
	}

	// Check that a solution exists.
	// if len(sols) > 1 {
	// 	panic("no solution exists")
	// }
	for i := range ns_ {
		for j := range ns_ {
			gcd := mathutil.GCD(ns_[i], ns_[j])
			if as_[i]%gcd != as_[j]%gcd {
				panic("no solution exists")
			}
		}
	}

	// Remove equations of the form x = 0 (mod 1). These are redundant.
	for i := 0; i < len(ns_); i++ {
		if ns_[i] == 1 {
			as_ = append(as_[:i], as_[i+1:]...)
			ns_ = append(ns_[:i], ns_[i+1:]...)
		}
	}

outer:
	for {
		for i := range ns_ {
			for j := range ns_ {
				// Don't compare with self.
				if i == j {
					continue
				}
				// Maintain that ns[i] >= ns[j].
				if ns_[i] < ns_[j] {
					continue
				}

				// Ensure as are simplified
				as_[i] %= ns_[i]

				// Remove equations of the form x = 0 (mod 1). These are redundant.
				if ns_[j] == 1 {
					as_ = append(as_[:j], as_[j+1:]...)
					ns_ = append(ns_[:j], ns_[j+1:]...)
					continue outer
				}

				// Remove duplicate equations.
				if as_[i] == as_[j] && ns_[i] == ns_[j] {
					as_ = append(as_[:i], as_[i+1:]...)
					ns_ = append(ns_[:i], ns_[i+1:]...)
					continue outer
				}

				// Simplify coprime equations if possible.
				gcd := mathutil.GCD(ns_[i], ns_[j]) // e.g. gcd(74, 37) = 37

				if gcd != 1 {
					iMultiplicity := 1
					for ni := ns_[i]; ni > 0 && ni%gcd == 0; ni /= gcd {
						iMultiplicity++
					}
					jMultiplicity := 1
					for nj := ns_[j]; nj > 0 && nj%gcd == 0; nj /= gcd {
						jMultiplicity++
					}

					if iMultiplicity < jMultiplicity {
						i, j = j, i
					}

					// Check if possible.
					// E.g. [x=3(mod 37) & x=3(mod 74)] can be simplified to [x=3(mod 37) & x=1(mod 2)]
					// But [x=2(mod 37) & x=3(mod 74)] are impossible since 2 != 3 (mod 37).
					if as_[i]%gcd != as_[j]%gcd {
						msg := fmt.Sprintf("x cannot simultaneously be %d (mod %d) and %d (mod %d)", as_[i], ns_[i], as_[j], ns_[j])
						panic(msg)
					}

					ns_[j] /= gcd
					as_[j] %= ns_[j]
					continue outer
				}
			}
		}
		break
	}

	return as_, ns_
}
